Revert "refactor(settings): move headers to include subdir"
This reverts commit 6dea43b248de5d94318a8025567f91a192057e23.
