+++ plugin/src/Caelestia/Settings/changebatcher.hpp
