+++ plugin/src/Caelestia/Settings/common.cpp
@@ -1,6 +1,6 @@
-#include "include/common.hpp"
+#include "common.hpp"
 
-#include "include/node.hpp"
+#include "node.hpp"
 
 namespace {
