+++ plugin/src/Caelestia/Settings/common.hpp
