+++ plugin/src/Caelestia/Settings/layerregistry.hpp
