+++ plugin/src/Caelestia/Settings/macros.hpp
