+++ plugin/src/Caelestia/Settings/objectnode.hpp
