+++ plugin/src/Caelestia/Settings/quarantine.cpp
@@ -1,4 +1,4 @@
-#include "include/quarantine.hpp"
+#include "quarantine.hpp"
 
 #include <qjsonobject.h>
