+++ plugin/src/Caelestia/Settings/quarantine.hpp
