+++ plugin/src/Caelestia/Settings/rootnode.cpp
@@ -1,4 +1,4 @@
-#include "include/rootnode.hpp"
+#include "rootnode.hpp"
 
 #include <qjsonobject.h>
