+++ plugin/src/Caelestia/Settings/schema.hpp
