+++ plugin/src/Caelestia/Settings/settingsfile.cpp
@@ -1,4 +1,4 @@
-#include "include/settingsfile.hpp"
+#include "settingsfile.hpp"
 
 #include <qdir.h>
 #include <qfile.h>
